Move DiscountTag from OfferNew to own component
